Feat: red outline, red links on empty player name
Highlighting is active while the player name is still unset, oder
whenever an existing entry in the input field is cleared out again at a
later point.

This encompassed turning the "interactive" links in the pv section into a
template so as to reduce the amount of copy-pasting Alpine code needed.
(i.e., light refactoring)